Title: 앞의 두 수를 더한 수열 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240321/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.java
@@ -6,7 +6,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int N = sc.nextInt();
 
-        int[] dp = new int[N];
+        long[] dp = new long[N];
         dp[0] = dp[1] = 1;
 
         for(int i=2; i<N; i++)
